# src/swagger/swaggerApi2.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Movie",
    "version": "2.0.0",
    "description": "API RESTful pour la gestion des utilisateurs, des films  et l’authentification sécurisée"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v2"
    }
  ],
  "paths": {
    "/users": {
      "get": {
        "summary": "Lister les utilisateurs",
        "tags": ["Users"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des utilisateurs"
          }
        }
      },
      "post": {
        "summary": "Créer un nouvel utilisateur",
        "tags": ["Users"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Utilisateur créé"
          },
          "400": {
            "description": "Erreur de requête"
          }
        }
      }
    },
    "/login": {
      "post": {
        "summary": "Authentifier un utilisateur",
        "tags": ["Users"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLogin"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authentification réussie"
          },
          "401": {
            "description": "Échec de l'authentification"
          }
        }
      }
    },
  "/me": {
        "get": {
          "summary": " afficher l'utilisateur connecté",
          "tags": ["Users"],
          
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "responses": {
            "200": {
              "description": "affiche utilisateur connecté"
            }
          }
        },
        "patch": {
          "summary": "Mettre à jour les informations de l'utilisateur connecté",
          "tags": ["Users"],
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserUpdate"
                }
              }
            }
          }
      }
  },
...
},
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "UserCreate": {
        "type": "object",
        "properties": {
          "nom": { "type": "string", "example": "bob" },
          "username": { "type": "string", "example": "alice42" },
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "minLength": 6, "example" : "password1!"},
          "role": { "type": "string", "example": "user" },
          "favorites": { "type": "string", "example": "64kw132ksk" }
        },
        "required": ["nom", "username", "email", "password", "role" , "favorites"]
      },
      "UserLogin": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string" }
        },
        "required": ["email", "password"]
      },
      "UserUpdate": {
        "type": "object",
        "properties": {
          "username": { "type": "string", "example": "alice42" },
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "minLength": 6 , "example" : "password1!" }
        }
      
      },
...
    }
  }
}